Fixes test/make_tp_collect_fixture_x series.
Series had not been updated to work the same way the tests work, which
is to only use two text files copied into the top level of
test/fixtures, rather than requiring a whole bloated project dir
fixture with mp3 files.

Moves a method from test_integration_scripts_3_tp_collect.rb into
Audibleturk::Test so it can be re-used in one of the make scripts.

MAKE SCRIPT CHANGES ARE UNTESTED. We have a working fixture and I'm
not going to fuck with making a new one for the moment. Big pain in
the ass. TODO.