bin/tp-make can re-try to upload a project that failed or partially failed to upload on creation.
tp-make will abort on failed upload. But if you call it again with the
same arguments, it will attempt a re-upload on the project.

Moved working_url? and fetch_url into Utility from Test to support
tp-make. Changed Project#create_assignment_csv to support a field that
indicates if an uplaod had been confirmed or not. Updated tests to
test for re-trying failed uploads.