Aborts when the URLs in Project#local#file('data','assignment.csv') do not match the URL prefixes specified in the current Config.
Specifically, Project#remote#url_basename raises an exception if it
cannot split an URL on the URL prefix in the config file. Also,
App.updelete_assignment_assets and App.upload_audio_for_project will
catch this exception and abort in a friendly manner that prompts the
user to make sure he is passing the correct --config arg.

This is to handle cases where the user passes in one --config path
while making/assigning a project but fails (forgets) to pass in that
same config when working on it subsequently.