Fixes bug sorting multiple input files on Linux.
Audio files submitted to tp-make were not being explicitly sorted
before merge, and were thus merged in whatever order they were in when
returned by the OS file listing. On Linux, this listing was (at least)
sometimes not in the order specified in the documentation (a simple
sort, per Array#sort in ruby). We now explicitly sort the input files
as specified in the documentation. (Perhaps it would be better to sort
in the order submitted by the user in args to tp-make.)