Fixes critical bug throwing error when encountering HITs with non external questions.
Adds regression test to test/test_unit_amazon.rb. Adds fixture for
regression test.

When looking for external question data in the HIT XML,
Typingpool::Amazon::HIT::Full was stupidly failing to handle the
possibility the XML might not contain an external question but instead
a QuestionForm, for example because it is a HIT not created by
Typingpool. It looks like I introduced this bug when switching
Typingpool away from QuestionForms last year.

Thanks to @bartramnason for the bug report.