Speeds up integration tests.
Instead of re-merging and re-splitting mp3 files over and over with
calls to tp-make, integration tests often now simply copy over a
ready-made project folder, which is generated at the start of the test
run and then deleted at the very end.

Adds a unit test to test this element of the test class.