Fixes test bug.
Prior code failed to reconfigure fixture assignment.csv URLs in
tp-assign tests when copying fixtures over, resulting in an error if
you tried to run the test with an S3 bucket with a different name than
the one used to record it. This fix should make the tests run on
arbitrary machines again.