Simplifies test.
There's been a longstanding problem with the bundled audio files: They
merge into exactly 2 minutes and we split on 20 seconds. Depending on
the vagaries of which mp3splt and mp3wrap version you use and what
platform you're running on, this split could result in 6 or 7
files. Previously this was no big deal because we'd just adjust
certain tests to work with either six OR seven audio files, most did
not care.

But with the bundling of extensive VCR mocks this has become an issue:
We create mocks based on a run of six audio files (uploading,
downloading, assigning HITs, etc). If these mocks are run in an
environment where the original audio splits into seven rather than six
files, the mocks don't cover all possible network calls because now
there is an extra file to upload, assign, delete, etc.

The solution was simple: Change the split time from 20 seconds to 22
seconds, producing a uniform six chunks not suseptible to
vagaries. This has produced remarkably few roadbumps.