Greatly reduced total file size.
Packaged as a gem, v0.8.6 of Typingpool was nearly 25MB, an
irresponsible amount of data to be placing on rubygems.org. This
version is 3.7MB.

The slimming is due to no longer saving base64 encoded versions of mp3
uploads in our VCR fixtures (fixtures that multiplied greatly in
v0.8.6). Now we filter out large upload bodies from these fixtures,
with no loss of functionality.